+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Since last release
 * Consistently use hyphens in ``install.py`` flags (#1748)
 * Material sell policy can package materials (#1749, #1774)
 * Use miniforge for conda CI builds instead of miniconda (#1763)
+* Define constants ``CY_LARGE_DOUBLE``, ``CY_LARGE_INT``, and ``CY_NEAR_ZERO`` (#1757)
 * Warning and limits on number of packages that can be created from a resource at once (#1771)
 
 **Removed:**